+++ heuristic.go
@@ -2,7 +2,7 @@ package main
 
 import "fmt"
 
-func hamming(puzMap MyMap, env *Env) (h float64) {
+func hamming(puzMap MyMap, env *Env) (h float32) {
 	goalMap := env.goal.puzMap
 	for i := 0; i < env.longSize; i++ {
 		if puzMap[i] != goalMap[i] {
@@ -12,7 +12,7 @@ func hamming(puzMap MyMap, env *Env) (h float64) {
 	return
 }
 
-func manhattan(puzMap MyMap, env *Env) (h float64) {
+func manhattan(puzMap MyMap, env *Env) (h float32) {
 	goalMap := env.goal.puzMap
 
 	for i := 0; i < env.longSize; i++ {
@@ -21,7 +21,7 @@ func manhattan(puzMap MyMap, env *Env) (h float64) {
 			for j := i; j < env.longSize && puzMap[i] != goalMap[j]; j++ {
 				tmp++
 				if puzMap[i] == goalMap[j] {
-					h += ((float64)(tmp/env.size) * (float64)(tmp/env.size)) + ((float64)(tmp%env.size) * (float64)(tmp%env.size))
+					h += ((float32)(tmp/env.size) * (float32)(tmp/env.size)) + ((float32)(tmp%env.size) * (float32)(tmp%env.size))
 				}
 			}
 		}
@@ -29,7 +29,7 @@ func manhattan(puzMap MyMap, env *Env) (h float64) {
 	return
 }
 
-func invCount(puzMap MyMap, env *Env) (h float64) {
+func invCount(puzMap MyMap, env *Env) (h float32) {
 	fmt.Println(puzMap)
 	for i := 0; i < env.longSize-1; i++ {
 		for j := i + 1; j < env.longSize && (puzMap[i] == env.goal.puzMap[j]); j++ {
